🔨 : profile modal 컴포넌트를 제거함
- 프로필 모달 컴포넌트를 제거하고, Modal 컴포넌트에 추가함
