--- front-end/src/components/ProfileModal/style.scss
@@ -1,78 +1,0 @@
-@import 'common/styles/base';
-
-@mixin profileModal--root {
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  position: fixed;
-}
-
-.profile-modal__overlay {
-  @include profileModal--root;
-  background: rgba(0, 0, 0, 0.4);
-}
-
-.profile-modal__root {
-  @include profileModal--root;
-
-  .profile-modal__content {
-    cursor: default;
-    position: absolute;
-    top: 45%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    line-height: 1.4;
-    //background: $transparent-light-navy;
-    background: rgba(54, 65, 113, 0.7);
-    padding: 20px 28px;
-    border: 1px solid white;
-    max-width: 600px;
-    min-width: 300px;
-    display: flex;
-    flex-direction: column;
-    color: white;
-    align-items: center;
-    font-size: 18px;
-
-    .profile-modal__header {
-      width: 100%;
-      display: flex;
-      justify-content: space-between;
-
-      .profile-modal__exit {
-        cursor: pointer;
-      }
-    }
-
-    .profile-modal__image {
-      width: 93px;
-      margin: 20px auto 15px;
-    }
-
-    .profile-modal__more {
-      cursor: pointer;
-      font-size: 16px;
-    }
-
-    .profile-modal__status {
-      width: 220px;
-      background-color: rgba(255, 255, 255, 0.15);
-      border: 1px rgba(255, 255, 255, 0.6) solid;
-      height: 120px;
-      border-radius: 5px;
-      font-size: 16px;
-      margin: 20px auto;
-      box-sizing: border-box;
-      padding: 10px;
-      color: white;
-      -ms-overflow-style: none;
-      scrollbar-width: none;
-      &::-webkit-scrollbar {
-        display: none;
-      }
-    }
-  }
-}
